Implement OpenID Connect authorization code grant_type
Note: this is a standalone implementation of [0]. The code was on hold
until further discussions in how to improve OpenID Connect support take
place.

[0] https://review.openstack.org/#/c/330006/

What follows is the original commit message:

---8<---

The current code is not working appropriately, as the authorization code
is a short-lived (the specification recommends less than 10 minutes) and
single use code, so it cannot be reused (if reused, the specification
states that all tokens obtained with the original code should be
revoked). Therefore, if we are implementing the authorization code grant
type we should follow the complete flow, that is:

1. obtain an authz code from the authz endpoint.
2. exchange the authz code with the token endpoint to get an access
   token.
3. exchange the access token with keystone to obtain an keystone token.

However (1) requires that the user authenticates with the IdP, and
According to the OAUTH 2.0 specification [1]:

    The authorization code grant type is used to obtain both access
    tokens and refresh tokens and is optimized for confidential clients.
    Since this is a redirection-based flow, the client must be capable
    of interacting with the resource owner’s user-agent (typically a web
    browser) and capable of receiving incoming requests (via
    redirection) from the authorization server.

[1] https://tools.ietf.org/html/rfc6749#section-4.1

So interacting with the user browser is expected. If it is not possible
to interact with the client in this way, then a different grant type
(like client_credentials (to be implemented)) should be used if
supported by the server.

In short, whenever we execute this authentication we need to fire up a
user-agent (browser) and we need to spawn a temporary local http server
where the request will be redirected so that we can get the returned
authorization code. Then, we can continue with the flow (i.e. obtaining
the access token and exchanging it with Keystone for a Keystone token).

Incidentally, the OpenID plugin does not accept the redirect_uri
argument anymore, as this is something internally managed by us, and
that the user should not provide.

Note that with the code as it is, the access token is not persisted, so
whenever the session ends it is lost. If a user makes 100 different
requests this implies that 100 browser windows or tabs will be opened.
A possible improvement for all OpenID Connect plugins is the persistence
of the access token so that it can be reused, and the implementation of
refresh tokens, instead of requiring authentication each time.

Closes-Bug: #1583961
Closes-Bug: #1496258
Change-Id: I154b1d4a45a8a8272f286895afa1c14ff7044cf4

--->8---